+++ tests/Properties/JsonProperties.hs
@@ -48,5 +48,6 @@ jsonProps = testGroup "JSON Parsers"
     , genProp "Errors"           (prop_json :: Errors -> Bool)
     , genProp "System"           (prop_json :: System -> Bool)
     , genProp "DirTree"          (prop_json :: DirTree -> Bool)
+    , genProp "UsageReport"      (prop_json :: UsageReport -> Bool)
     ]